+++ lib/index.ts
