+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
         "declaration": true,
-        "outDir": "./lib",
+        "outDir": "./out",
         "target": "es2015",
         "module": "commonjs",
         "strict": true,
@@ -10,5 +10,5 @@
             "types"
         ]
     },
-    "include": ["./src/**/*"]
+    "include": ["lib", "test"]
 }
